legacy/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
deprecation==2.1.0
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
packaging==24.2
postgrest==0.18.0
propcache==0.2.1
psutil==5.9.7
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-magic==0.4.27
python-telegram-bot==21.9
realtime==2.0.6
redis==5.0.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
yarl==1.18.3
zlib_wrapper==0.1.3
python-telegram-bot[job-queue]==21.9
watchdog==6.0.0
beautifulsoup4==4.12.3
googlesearch-python==1.2.5
trafilatura==2.0.0
google-generativeai==0.8.3
Pillow==11.0.0
sentence_transformers==3.3.1